electron.remote is remonved since electron version 14. All reference to electron.remote is comment out to upgrade the Biglots OpenPOS to the current electron version. (#2300)
biglots defect: TS-12785

+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/services/session.service.ts
@@ -206,18 +206,22 @@ export class SessionService implements IMessageHandler<any> {
      * Need to come up with a better way to enapsulate electron and node ... should put these reference behind our new platform interface
      */
     private deleteLaunchingFlg() {
-        const fs = this.electron.isElectronApp ? this.electron.remote.require('fs') : window.fs;
-        const launchingFile = 'launching.flg';
-        this.log.info('node.js fs exists? ' + fs);
-        this.log.info('launching.flg file exists? ' + (fs && fs.existsSync(launchingFile)));
-        if (fs && fs.existsSync(launchingFile)) {
-            fs.unlink(launchingFile, (err) => {
-                if (err) {
-                    this.log.info('unable to remove ' + launchingFile);
-                } else {
-                    this.log.info(launchingFile + ' was removed');
-                }
-            });
+        /*const fs = this.electron.isElectronApp ? this.electron.remote.require('fs') : window.fs;*/
+        if (this.electron.isElectronApp) {
+        } else {
+            const fs = window.fs;
+            const launchingFile = 'launching.flg';
+            this.log.info('node.js fs exists? ' + fs);
+            this.log.info('launching.flg file exists? ' + (fs && fs.existsSync(launchingFile)));
+            if (fs && fs.existsSync(launchingFile)) {
+                fs.unlink(launchingFile, (err) => {
+                    if (err) {
+                        this.log.info('unable to remove ' + launchingFile);
+                    } else {
+                        this.log.info(launchingFile + ' was removed');
+                    }
+                });
+            }
         }
     }
